Stop eagerly importing the config-template stack from model and objectstore
Follow-up to the set_metadata import work. This one is mostly structural: the
measured gain is small, 1.89s -> 1.85s per metadata process (about 2 percent,
5 fewer modules), because the same modules are still reached through paths that
are genuinely required. It is committed for the coupling it removes, not as a
performance claim.

- galaxy.objectstore imported .templates purely for an annotation. The module
  already uses "from __future__ import annotations", so the import moves to
  TYPE_CHECKING unchanged. User-defined object store templates are not needed to
  operate an object store.
- galaxy.model imported galaxy.files.templates and galaxy.objectstore.templates
  at module level, but every use is inside two methods of the UserFileSource and
  UserObjectStore classes. Deferred to those call sites.
- partial_model and its helpers lived in galaxy.util.config_templates but are
  generic pydantic utilities with no connection to configuration templates.
  Importing them from galaxy.schema.schema meant galaxy.model pulled in
  config_templates and, transitively, requests, yaml, boltons and logfire. Moved
  to galaxy.util.pydantic_partial; config_templates keeps its own uses.

Why the gain is only 2 percent: galaxy.schema.schema is required by galaxy.model
for the state enums and independently imports galaxy.tool_util_models, and
galaxy.datatypes.sniff is required by set_metadata and independently imports
galaxy.files. Cutting the redundant paths therefore does not remove the modules,
only the surplus edges. The remaining ~1.85s is dominated by work that cannot be
deferred: SQLAlchemy declarative mapping in galaxy.model (~0.6s) and pydantic
model construction in galaxy.schema.schema and galaxy.tool_util_models (~0.4s).
