+++ lib/galaxy/util/config_templates.py
@@ -6,12 +6,9 @@ This is capturing code shared by file source templates and object store template
 import logging
 import os
 from collections.abc import (
-    Callable,
-    Iterable,
     Sequence,
 )
 from typing import (
-    Annotated,
     Any,
     cast,
     Literal,
@@ -26,11 +23,9 @@ from boltons.iterutils import remap
 from pydantic import (
     BaseModel,
     ConfigDict,
-    create_model,
     RootModel,
     ValidationError,
 )
-from pydantic.fields import FieldInfo
 from typing_extensions import (
     NotRequired,
     Protocol,
@@ -704,50 +699,3 @@ class ImplicitConfigurationParameters(TypedDict):
     oauth2_access_token: NotRequired[str]
 
 
-M = TypeVar("M", bound="BaseModel")
-
-
-# Implementation copied from https://github.com/pydantic/pydantic/issues/12329#issuecomment-3382159312
-def _make_field_optional(field_info: FieldInfo):
-    """Returns the field's definition to be used in a `create_model()` call to make the field optional."""
-    annotation = field_info.annotation
-    assert annotation is not None
-    if field_info.is_required():
-        return Annotated[annotation | None, field_info], None
-    else:
-        return Annotated[annotation, field_info]
-
-
-def make_model_with_all_fields_optional(model: type[M], fields=None) -> type[M]:
-    """Returns a new Pydantic model based on `model`, but with all fields optional."""
-    if fields is None:
-        fields = model.model_fields.items()
-    return create_model(
-        model.__name__,
-        __doc__=model.__doc__,
-        __base__=model,
-        **{field_name: _make_field_optional(field_info) for field_name, field_info in fields},
-    )
-
-
-# TODO: This is a workaround to make all fields optional.
-#       It should be removed when Python/pydantic supports this feature natively.
-# https://github.com/pydantic/pydantic/issues/1673
-def partial_model(include: list[str] | None = None, exclude: list[str] | None = None) -> Callable[[type[M]], type[M]]:
-    """Decorator to make all model fields optional"""
-
-    if exclude is None:
-        exclude = []
-
-    def decorator(model: type[M]) -> type[M]:
-        if include is None:
-            fields: Iterable[tuple[str, FieldInfo]] = model.model_fields.items()
-        else:
-            fields = ((k, v) for k, v in model.model_fields.items() if k in include)
-
-        if exclude is not None:
-            fields = ((k, v) for k, v in fields if k not in exclude)
-
-        return make_model_with_all_fields_optional(model, fields)
-
-    return decorator
